fix(pci): fix(pci): get the flavors from the api for each instance
The initialization of the instance is not necessarily complete after the promise of init
infrastructure is done, because the internal promises are not well managed. As a result, sometimes
the flavors are not initialized in time. Flavor id value should always be available even if all
promises are not not completed.